+++ dapps/src/apps/mod.rs
@@ -20,10 +20,8 @@ use proxypac::ProxyPac;
 use parity_dapps::WebApp;
 
 mod fs;
-mod fetcher;
+pub mod fetcher;
 pub mod manifest;
-
-pub use self::fetcher::AppFetcher;
 
 extern crate parity_dapps_status;
 extern crate parity_dapps_home;
