+++ dapps/src/handlers/client/fetch_file.rs
