+++ dapps/src/lib.rs
@@ -124,7 +124,7 @@ impl Server {
 	fn start_http<A: Authorization + 'static>(addr: &SocketAddr, authorization: A, handler: Arc<IoHandler>, dapps_path: String) -> Result<Server, ServerError> {
 		let panic_handler = Arc::new(Mutex::new(None));
 		let authorization = Arc::new(authorization);
-		let apps_fetcher = Arc::new(apps::AppFetcher::new(&dapps_path));
+		let apps_fetcher = Arc::new(apps::fetcher::AppFetcher::new(&dapps_path));
 		let endpoints = Arc::new(apps::all_endpoints(dapps_path));
 		let special = Arc::new({
 			let mut special = HashMap::new();
