+++ dapps/src/router/mod.rs
@@ -26,7 +26,8 @@ use std::collections::HashMap;
 use url::{Url, Host};
 use hyper::{self, server, Next, Encoder, Decoder, Control};
 use hyper::net::HttpStream;
-use apps::{self, AppFetcher};
+use apps;
+use apps::fetcher::AppFetcher;
 use endpoint::{Endpoint, Endpoints, EndpointPath};
 use handlers::{Redirection, extract_url};
 use self::auth::{Authorization, Authorized};
